Inline fetch functions in pixman-transformed.c
There was significant performance overhead associated with indirect
call to the fetch functions, so inline them.

Also, the old code assumed that if the number of rectangles in the
source clip was 1, then the clip was identical to the image
boundaries. Fix that by running a full region_contains_point()
whenever the source clip is different from pict->common.full_region.

Based on a patch from Antoine Azar.