Rename the current ARM code to ARM SIMD
This code is only for CPUs supporting the SIMD instructions, not for all ARM
CPUs.

I stumbled above the recent commit with the ARM SIMD code while preparing a
patch that models the patch from #13445 after the MMX and SSE2 cases:

The ARM SIMD option currently uses --disable-arm, although this code is only
for CPUs >= ARMv6. That's as if one would call the option to disable the SSE2
code --disable-x86.

This patch therefore renames the configure option and the function and file
names to arm-simd/arm_simd.