Turn off asserts in development snapshots (bug 26314).
There is not much real benefit in having asserts turned on in
snapshots because it doesn't lead to any new bug reports, just to
people not installing development snapshots since they case X server
crashes. So just turn them off.

While we are at it, limit the number of messages to stderr to 5
instead of 50.